+++ RELEASENOTES.md
@@ -8,6 +8,9 @@ mulle-builds sake)
 * Because `mulle-bootstrap setting` now works, init by default does not
 create demo files any more. Inverted meaning of -d flag.
 * Various improvements to mulle-brew handling
+* The paths default is now to output an unquoted one-liner. This is less
+correct but simpler.
+
 
 ### 3.4.0